requirements-test.txt
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
PyYAML==6.0.1